+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 
 - Common
     - Set `probe_id` as `probe_description` when inserting from nwb file #1220
+    - Default `AnalysisNwbfile.create` permissions are now 777 #1226
 - Position
     - Allow population of missing `PositionIntervalMap` entries during population
         of `DLCPoseEstimation` #1208